+++ haulage_mgmt/haulage_logistics/doctype/haulage_trip_custody/haulage_trip_custody.py
@@ -1,0 +1,5 @@
+from frappe.model.document import Document
+
+
+class HaulageTripCustody(Document):
+    pass